Convert prefix to a keyword argument
This is a breaking change for callers, but needed so that I can allow
multiple codes to be passed so we have have custom combinations of
markers all on one page. This is potentially useful so we could
easily print a bunch of sample markers of a given kind on a single
sheet.

Change-Id: I203fde74420f929cea6f135899b4160e1ad5303f