Add a script which validates the generated markers
This generates images for the first 200 markers, then runs libkoki
on the generated files and checks whether the correct value is
returned. It would be nice not to have to go via files, but OpenCV
doesn't appear to have a nice way to do this.

This currently shows upa small number of failures.

Change-Id: I1a03e68f60729df12dcc400e856b1ba480d4e489